Changing the variable type from byte to long. It will avoid troubles between intents.

+++ src/org/ricciardelli/shopping/CRUD.java
@@ -15,7 +15,7 @@
 package org.ricciardelli.shopping;
 
 public interface CRUD {
-	void create(byte id);
+	void create(long id);
 
 	void read(long id);
